+++ db/migrate/20191015122729_add_confirmation_report_uploaded_by_to_payroll_runs.rb
@@ -1,0 +1,5 @@
+class AddConfirmationReportUploadedByToPayrollRuns < ActiveRecord::Migration[5.2]
+  def change
+    add_column :payroll_runs, :confirmation_report_uploaded_by, :string
+  end
+end